+++ settings.php
@@ -71,6 +71,81 @@ if ($hassiteconfig) {
                 get_string('keepcourseid_desc', 'local_sandbox', null, true),
                 0));
 
+        // Restore settings.
+        $html = html_writer::tag('p', get_string('restoresettingsdescription', 'local_sandbox', null, true));
+        $html .= html_writer::tag('p', get_string('restoresettingswarning', 'local_sandbox', null, true));
+        $page->add(new admin_setting_heading('local_sandbox/restoreheading',
+                get_string('restoreheading', 'local_sandbox', null, true),
+                $html));
+
+        // Create restore settings analogous to the Moodle core settings on /admin/settings/courses.php.
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_users',
+                get_string('generalusers', 'backup', null, true),
+                get_string('configrestoreusers', 'backup', null, true),
+                0));
+        $options = [
+                // Can not use actual constants here because we'd need to include 100 of backup/restore files.
+                // Don't use string lazy loading here because the strings will be directly used and
+                // would produce a PHP warning otherwise.
+                0/*backup::ENROL_NEVER*/     => get_string('rootsettingenrolments_never', 'backup'),
+                1/*backup::ENROL_WITHUSERS*/ => get_string('rootsettingenrolments_withusers', 'backup'),
+                2/*backup::ENROL_ALWAYS*/    => get_string('rootsettingenrolments_always', 'backup'),
+        ];
+        $page->add(new admin_setting_configselect('local_sandbox/restore_general_enrolments',
+                get_string('generalenrolments', 'backup', null, true),
+                get_string('configrestoreenrolments', 'backup', null, true),
+                0,
+                $options));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_role_assignments',
+                get_string('generalroleassignments', 'backup', null, true),
+                get_string('configrestoreroleassignments', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_activities',
+                get_string('generalactivities', 'backup', null, true),
+                get_string('configrestoreactivities', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_blocks',
+                get_string('generalblocks', 'backup', null, true),
+                get_string('configrestoreblocks', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_filters',
+                get_string('generalfilters', 'backup', null, true),
+                get_string('configrestorefilters', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_comments',
+                get_string('generalcomments', 'backup', null, true),
+                get_string('configrestorecomments', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_badges',
+                get_string('generalbadges', 'backup', null, true),
+                get_string('configrestorebadges', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_calendarevents',
+                get_string('generalcalendarevents', 'backup', null, true),
+                get_string('configrestorecalendarevents', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_userscompletion',
+                get_string('generaluserscompletion', 'backup', null, true),
+                get_string('configrestoreuserscompletion', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_logs',
+                get_string('generallogs', 'backup', null, true),
+                get_string('configrestorelogs', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_histories',
+                get_string('generalhistories', 'backup', null, true),
+                get_string('configrestorehistories', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_groups',
+                get_string('generalgroups', 'backup', null, true),
+                get_string('configrestoregroups', 'backup', null, true),
+                0));
+        $page->add(new admin_setting_configcheckbox('local_sandbox/restore_general_competencies',
+                get_string('generalcompetencies', 'backup', null, true),
+                get_string('configrestorecompetencies', 'backup', null, true),
+                0));
+
+
         // Notifications.
         $page->add(new admin_setting_heading('local_sandbox/notifyheading',
                 get_string('notifyheading', 'local_sandbox', null, true),